requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
llama-cpp-python>=0.2.0
pydantic>=1.10.0
